chg:usr:Get query stats for a certain timeperiod
Previously query stats were calculated for all queries. This is repetitive
and might be costly. The queries have been changed to only calculate for the
last X minutes where X is the frequency of the cron.
The range is on endTime of the query as queries are added only when it is complete.